+++ api/src/components/publication/service.ts
@@ -355,66 +355,6 @@ export const updateStatus = async (id: string, status: I.PublicationStatusEnum, 
     return updatedPublication;
 };
 
-export const createFlag = async (
-    publication: string,
-    user: string,
-    category: I.PublicationFlagCategoryEnum,
-    comment: string
-) => {
-    const flag = await client.prisma.publicationFlags.create({
-        data: {
-            category,
-            user: {
-                connect: {
-                    id: user
-                }
-            },
-            flagComments: {
-                create: {
-                    comment,
-                    createdBy: user
-                }
-            },
-            publication: {
-                connect: {
-                    id: publication
-                }
-            }
-        }
-    });
-
-    return flag;
-};
-
-export const getFlag = async (id: string) => {
-    const flag = await client.prisma.publicationFlags.findFirst({
-        where: {
-            id
-        },
-        include: {
-            publication: {
-                include: {
-                    user: true
-                }
-            }
-        }
-    });
-
-    return flag;
-};
-
-export const createFlagComment = async (id: string, comment: string, user: string) => {
-    const flagComment = await client.prisma.flagComments.create({
-        data: {
-            flagId: id,
-            comment,
-            createdBy: user
-        }
-    });
-
-    return flagComment;
-};
-
 export const validateConflictOfInterest = (publication: I.Publication) => {
     if (publication.conflictOfInterestStatus) {
         if (!publication.conflictOfInterestText?.length) return false;
@@ -460,17 +400,4 @@ export const isPublicationReadyToPublish = (publication: I.Publication, status: 
         isReady = true;
 
     return isReady;
-};
-
-export const resolveFlag = async (id: string) => {
-    const resolveFlag = await client.prisma.publicationFlags.update({
-        where: {
-            id
-        },
-        data: {
-            resolved: true
-        }
-    });
-
-    return resolveFlag;
 };
